+++ lib/features/loan_detail/presentation/widgets/payment_detail_card.dart
@@ -61,7 +61,7 @@ class PaymentDetailCard extends StatelessWidget {
                         ),
                         TextSpan(
                           text: Formatter.formatMoney(
-                            payment.paymentAmount.toDouble(),
+                            payment.paymentAmount,
                           ),
                           style: TextStyle(
                             color: Theme.of(context).colorScheme.onSurface,
